♻️ refactor : 입출력 기능에대한 모듈화
/src/input 폴더에  입출력 관련한 메서드들 모듈화 완성 및 App.js 코드 수정.

ex) Fixes : #45 Related to : #34, #23
