+++ src/input/index.js
@@ -1,0 +1,2 @@
+export { default as CarName } from './car-name.js';
+export { default as PlayNumber } from './play-number.js';
